test: Update email and close account viewport aware
